% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.properties.data.R
\name{add.properties.data}
\alias{add.properties.data}
\title{add.properties.data}
\usage{
add.properties.data(data, Ps = c(12, 88))
}
\arguments{
\item{data}{data}
}
\description{
Returns data with pPs added
}
\author{
Félicien Meunier
}
