% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.properties.R
\name{add.properties}
\alias{add.properties}
\title{add.properties}
\usage{
add.properties(models, x = c(12, 50, 88))
}
\arguments{
\item{models}{models}

\item{x}{some PLC}
}
\description{
Add 2 list of models
}
\author{
Félicien Meunier
}
