% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumnorm.R
\name{cumnorm.comp}
\alias{cumnorm.comp}
\title{cumnorm.comp}
\usage{
cumnorm.comp(data, m = -2, std = 1)
}
\arguments{
\item{data}{data.frame with water potential ("psi") and PLC ("PLC")}

\item{m}{mean (P50)}

\item{std}{standard deviation}
}
\description{
Returns RMSE
}
\author{
Félicien Meunier
}
