% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumnorm.R
\name{cumnorm}
\alias{cumnorm}
\title{cumnorm}
\usage{
cumnorm(psi, m = -2, std = 1)
}
\arguments{
\item{psi}{water potential}

\item{m}{mean (P50)}

\item{std}{standard deviation}
}
\description{
Returns cumnorm PLC
}
\author{
Félicien Meunier
}
