% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.model.params.R
\name{default.model.params}
\alias{default.model.params}
\title{default.model.params}
\usage{
default.model.params(fun.name)
}
\arguments{
\item{fun.name}{Function name}
}
\description{
Returns default model param
}
\author{
Félicien Meunier
}
