% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.model.R
\name{extract.model}
\alias{extract.model}
\title{extract.model}
\usage{
extract.model(all.models, pos)
}
\arguments{
\item{all.models}{all.models}

\item{pos}{pos}
}
\description{
Returns model
}
\author{
Félicien Meunier
}
