% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.NA.data.R
\name{fill.NA.data}
\alias{fill.NA.data}
\title{fill.NA.data}
\usage{
\method{fill}{NA.data}(data.in = data)
}
\arguments{
\item{data.in}{data.in}
}
\description{
Returns full data.frame
}
\author{
Félicien Meunier
}
