% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.best.model.R
\name{find.best.model}
\alias{find.best.model}
\title{find.best.model}
\usage{
find.best.model(All.models)
}
\arguments{
\item{data.test}{data}

\item{function.names}{function.names}
}
\description{
Returns best model
}
\author{
Félicien Meunier
}
