% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumnorm.R
\name{invert.cumnorm}
\alias{invert.cumnorm}
\title{invert.cumnorm}
\usage{
invert.cumnorm(x, m = -2, std = 1)
}
\arguments{
\item{x}{PLC}

\item{m}{mean (P50)}

\item{std}{standard deviation}
}
\description{
Returns inverse polynomial
}
\author{
Félicien Meunier
}
