% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial.R
\name{invert.polynomial}
\alias{invert.polynomial}
\title{invert.polynomial}
\usage{
invert.polynomial(x, a = 2, b = -2)
}
\arguments{
\item{x}{PLC}

\item{a}{a}

\item{b}{b (P50)}
}
\description{
Returns inverse polynomial
}
\author{
Félicien Meunier
}
