% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial2.R
\name{invert.polynomial2}
\alias{invert.polynomial2}
\title{invert.polynomial2}
\usage{
invert.polynomial2(x, a = 2, b = -2)
}
\arguments{
\item{x}{PLC}

\item{a}{a}

\item{b}{b (P50)}
}
\description{
Returns inverse polynomial2
}
\author{
Félicien Meunier
}
