% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoidal.R
\name{invert.sigmoidal}
\alias{invert.sigmoidal}
\title{invert.sigmoidal}
\usage{
invert.sigmoidal(x, a = 2, b = -2)
}
\arguments{
\item{x}{PLC}

\item{a}{a}

\item{b}{b (P50)}
}
\description{
Returns inverse sigmoidal
}
\author{
Félicien Meunier
}
