% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{invert.weibull}
\alias{invert.weibull}
\title{invert.weibull}
\usage{
invert.weibull(x, lambda = 2, k = 2)
}
\arguments{
\item{x}{PLC}

\item{lambda}{lambda}

\item{k}{k}
}
\description{
Returns inverse weibull
}
\author{
Félicien Meunier
}
