% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.sources.R
\name{merge.sources}
\alias{merge.sources}
\title{merge.sources}
\usage{
\method{merge}{sources}(data)
}
\arguments{
\item{data}{data}
}
\description{
Returns data with ax/p50 corrected
}
\author{
Félicien Meunier
}
