% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs2params.R
\name{obs2params}
\alias{obs2params}
\title{obs2params}
\usage{
obs2params(ax50, P50, funbest)
}
\arguments{
\item{ax50}{ax50}

\item{P50}{P50}

\item{funbest}{funbest}
}
\description{
Returns function parameters
}
\author{
Félicien Meunier
}
