% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.data.R
\name{opt.data}
\alias{opt.data}
\title{opt.data}
\usage{
opt.data(data = data, function.names = c("cumnorm", "weibull",
  "sigmoidal", "polynomial", "polynomial2"))
}
\arguments{
\item{data}{data}

\item{function.names}{function.names}
}
\description{
Returns list
}
\author{
Félicien Meunier
}
