% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.models.R
\name{plot.models}
\alias{plot.models}
\title{plot.models}
\usage{
\method{plot}{models}(models, add = TRUE, col = "black",
  highlight = FALSE)
}
\arguments{
\item{models}{models}
}
\description{
plot models
}
\author{
Félicien Meunier
}
