% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial.R
\name{polynomial}
\alias{polynomial}
\title{polynomial}
\usage{
polynomial(psi, a = 2, b = -2)
}
\arguments{
\item{psi}{water potential}

\item{a}{a}

\item{b}{b (P50)}
}
\description{
Returns polynomial PLC
}
\author{
Félicien Meunier
}
