% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial2.R
\name{polynomial2.comp}
\alias{polynomial2.comp}
\title{polynomial2.comp}
\usage{
polynomial2.comp(data, a = 2, b = -2)
}
\arguments{
\item{data}{data.frame with water potential ("psi") and PLC ("PLC")}

\item{a}{a}

\item{b}{b (P50)}
}
\description{
Returns RMSE
}
\author{
Félicien Meunier
}
