% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoidal.R
\name{sigmoidal}
\alias{sigmoidal}
\title{sigmoidal}
\usage{
sigmoidal(psi, a = 2, b = -2)
}
\arguments{
\item{psi}{water potential}

\item{a}{a}

\item{b}{b (P50)}
}
\description{
Returns sigmoidal PLC
}
\author{
Félicien Meunier
}
