% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumnorm.R
\name{slope.cumnorm}
\alias{slope.cumnorm}
\title{slope.cumnorm}
\usage{
slope.cumnorm(x, m = -2, std = 1)
}
\arguments{
\item{x}{PLC value}

\item{m}{mean (P50)}

\item{std}{standard deviation}
}
\description{
Returns slope of cumnorm
}
\author{
Félicien Meunier
}
