% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial.R
\name{slope.polynomial}
\alias{slope.polynomial}
\title{slope.polynomial}
\usage{
slope.polynomial(x, a = 2, b = -2)
}
\arguments{
\item{x}{PLC value}

\item{a}{a}

\item{b}{b (P50)}
}
\description{
Returns slope of polynomial
}
\author{
Félicien Meunier
}
