% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial2.R
\name{slope.polynomial2}
\alias{slope.polynomial2}
\title{slope.polynomial2}
\usage{
slope.polynomial2(x, a = 2, b = -2)
}
\arguments{
\item{x}{PLC value}

\item{a}{a}

\item{b}{b (P50)}
}
\description{
Returns slope of polynomial2
}
\author{
Félicien Meunier
}
