% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoidal.R, R/weibull.R
\name{slope.weibull}
\alias{slope.weibull}
\alias{slope.sigmoidal}
\title{slope.weibull}
\usage{
slope.sigmoidal(x, a = 2, b = -2)

slope.weibull(x, lambda = 2, k = 2)
}
\arguments{
\item{x}{PLC value}

\item{a}{a}

\item{b}{b (P50)}

\item{lambda}{lambda}

\item{k}{k}

\item{x}{PLC value}
}
\description{
Returns slope of weibull

Returns slope of weibull
}
\author{
Félicien Meunier

Félicien Meunier
}
