% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{weibul.comp}
\alias{weibul.comp}
\alias{weibull.comp}
\title{weibul.comp}
\usage{
weibull.comp(data, lambda = 2, k = 2)
}
\arguments{
\item{data}{data.frame with water potential ("psi") and PLC ("PLC")}

\item{lambda}{lambda}

\item{k}{k}
}
\description{
Returns RMSE
}
\author{
Félicien Meunier
}
