% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{weibull}
\alias{weibull}
\title{weibull}
\usage{
weibull(psi, lambda = 2, k = 2)
}
\arguments{
\item{psi}{water potential}

\item{lambda}{lambda}

\item{k}{k}
}
\description{
Returns Weibull PLC
}
\author{
Félicien Meunier
}
