% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibul.R
\name{invert.weibull}
\alias{invert.weibull}
\title{invert.weibull}
\usage{
invert.weibull(x, param)
}
\arguments{
\item{x}{PLC}

\item{param}{vector of lambda and k}
}
\description{
Returns inverse weibull
}
\author{
Félicien Meunier
}
