% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibul.R
\name{slope.weibull}
\alias{slope.weibull}
\title{slope.weibull}
\usage{
slope.weibull(x, param)
}
\arguments{
\item{x}{PLC value}

\item{param}{vector of lambda and k}
}
\description{
Returns slope of weibull
}
\author{
Félicien Meunier
}
