% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibul.R
\name{weibul.comp}
\alias{weibul.comp}
\title{weibul.comp}
\usage{
weibul.comp(data, param)
}
\arguments{
\item{data}{data.frame with water potential ("psi") and PLC ("PLC")}

\item{param}{vector of lambda and k}
}
\description{
Returns RMSE
}
\author{
Félicien Meunier
}
