% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibul.R
\name{weibull}
\alias{weibull}
\title{weibull}
\usage{
weibull(psi, param)
}
\arguments{
\item{psi}{water potential}

\item{param}{vector of lambda and k}
}
\description{
Returns Weibull PLC
}
\author{
Félicien Meunier
}
